Describe and move the ViewModel
